Update symmetric param
