+++ tests/plotting/image_cache/axes_geometry_source_rgb_scalars[True].png
